refactor: Refactor codes of BoardView Path
BoardView Path에서 params와 search를 가져오는 로직을 hook로 정의했습니다.
