+++ application/millage-client/src/hooks/board/index.ts
@@ -1,1 +1,2 @@
 export {default as useBoard} from './useBoard';
+export {default as useBoardViewPath} from './useBoardViewPath';
